+++ Cargo.toml
@@ -4,7 +4,6 @@ members = [
     "crates/*",
     "crates/btc_signer",
     "crates/btc_aggregator",
-    "crates/btc_signer_types",
 ]
 
 [workspace.package]
@@ -30,6 +29,9 @@ hex = "0.4"
 bincode = "1.3"
 
 frost-secp256k1 = "2.1.0"
+
+proptest = "1.7.0"
+futures = "0.3.31"
 
 rand_core = "0.6.4"
 serde_json = "1.0.143"
