+++ crates/btc_aggregator/Cargo.toml
@@ -14,6 +14,9 @@ uuid = { workspace = true }
 
 btc-signer = { path = "../btc_signer" }
 frost-secp256k1 = { workspace = true }
+futures = { workspace = true }
+rand = { workspace = true }
+proptest = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true, features = ["test-util"] }
