--- crates/btc_signer_types/Cargo.toml
@@ -1,7 +1,0 @@
-[package]
-name = "btc-signer-types"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-
