--- crates/btc_signer_types/src/lib.rs
@@ -1,1 +1,0 @@
-pub mod types;
