--- crates/btc_signer_types/src/types.rs
@@ -1,2 +1,0 @@
-
-
